tweak for misuse with values > 32-bit long
git-svn-id: https://svn.r-project.org/R/trunk@56667 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/memory.c
@@ -2015,14 +2015,14 @@ char *S_alloc(long nelem, int eltsize)
 
 char *S_realloc(char *p, long new, long old, int size)
 {
-    int nold;
+    size_t nold;
     char *q;
     /* shrinking is a no-op */
     if(new <= old) return p;
-    q = R_alloc(new, size);
-    nold = old * size;
+    q = R_alloc((size_t)new, size);
+    nold = (size_t)old * size;
     memcpy(q, p, nold);
-    memset(q + nold, 0, new*size - nold);
+    memset(q + nold, 0, (size_t)new*size - nold);
     return q;
 }
